Reorganized fixtures in e2e accounts
