+++ .github/workflows/cron-main-e2e.yml
@@ -6,7 +6,7 @@ on:
 jobs:
   build:
     runs-on: ubuntu-latest
-    timeout-minutes: 15
+    timeout-minutes: 20
     steps:
     - name: "Checkout"
       uses: actions/checkout@v4
