+++ .github/workflows/pr-build-merge.yml
@@ -15,7 +15,7 @@ jobs:
   build:
 
     runs-on: ubuntu-latest
-    timeout-minutes: 15
+    timeout-minutes: 20
 
     steps:
     - name: "Checkout"
