+++ tests/e2e/accounts/buyers/conftest.py
@@ -1,0 +1,39 @@
+import pytest
+
+
+@pytest.fixture
+def invalid_buyer_id():
+    return "BUY-0000-0000"
+
+
+@pytest.fixture
+def buyer_account_id(e2e_config):
+    return e2e_config["accounts.buyer.account.id"]
+
+
+@pytest.fixture
+def buyer_factory(buyer_account_id):
+    def _buyer(
+        name="E2E Created Buyer",
+        account_id: str = buyer_account_id,
+    ):
+        return {
+            "name": name,
+            "account": {
+                "id": account_id,
+            },
+            "contact": {
+                "firstName": "first",
+                "lastName": "last",
+                "email": "created.buyer@example.com",
+            },
+            "address": {
+                "addressLine1": "123 Main St",
+                "city": "Anytown",
+                "state": "CA",
+                "postCode": "12345",
+                "country": "US",
+            },
+        }
+
+    return _buyer
