+++ tests/e2e/accounts/conftest.py
@@ -19,79 +19,33 @@ def currencies():
 
 
 @pytest.fixture
-def seller_factory(currencies):
-    def _seller(
-        external_id: str,  # Must be unique in Marketplace
-        name="E2E Test Seller",
-        currencies=currencies,
-    ):
-        return {
-            "name": name,
-            "address": {
-                "addressLine1": "123 Main St",
-                "city": "Anytown",
-                "state": "CA",
-                "postCode": "12345",
-                "country": "US",
-            },
-            "currencies": currencies,
-            "externalId": external_id,
-        }
-
-    return _seller
+def account_id(e2e_config):
+    return e2e_config["accounts.account.id"]
 
 
 @pytest.fixture
-def account_factory():
-    def _account(
-        name: str = "E2E Test Api Client Vendor",
-    ):
-        return {
-            "name": name,
-            "address": {
-                "addressLine1": "123 Test St",
-                "city": "San Francisco",
-                "state": "CA",
-                "postCode": "12345",
-                "country": "US",
-            },
-            "type": "Vendor",
-            "status": "Active",
-        }
-
-    return _account
+def seller_id(e2e_config):
+    return e2e_config["accounts.seller.id"]
 
 
 @pytest.fixture
-def buyer_factory(buyer_account_id):
-    def _buyer(
-        name="E2E Created Buyer",
-        account_id: str = buyer_account_id,
-    ):
-        return {
-            "name": name,
-            "account": {
-                "id": account_id,
-            },
-            "contact": {
-                "firstName": "first",
-                "lastName": "last",
-                "email": "created.buyer@example.com",
-            },
-            "address": {
-                "addressLine1": "123 Main St",
-                "city": "Anytown",
-                "state": "CA",
-                "postCode": "12345",
-                "country": "US",
-            },
-        }
+def buyer_id(e2e_config):
+    return e2e_config["accounts.buyer.id"]
 
-    return _buyer
+
+@pytest.fixture
+def user_group_id(e2e_config):
+    return e2e_config["accounts.user_group.id"]
+
+
+@pytest.fixture
+def module_id(e2e_config):
+    return e2e_config["accounts.module.id"]
 
 
 @pytest.fixture
 def user_group_factory(account_id, module_id):
+    # Used in user group and licensee fixtures
     def _user_group(
         name: str = "E2E Test Api Client User Group",
         user_group_account_id: str = account_id,
@@ -106,51 +60,3 @@ def user_group_factory(account_id, module_id):
         }
 
     return _user_group
-
-
-@pytest.fixture
-def api_token_factory(account_id, module_id):
-    def _api_token(
-        name: str = "E2E Test API Token",
-        description: str = "E2E API Token created during E2E tests",
-    ):
-        return {
-            "account": {"id": account_id},
-            "name": name,
-            "description": description,
-            "icon": "",
-            "modules": [{"id": module_id}],
-        }
-
-    return _api_token
-
-
-@pytest.fixture
-def licensee_factory(seller_id, buyer_id, user_group_factory, licensee_account_id):
-    def _licensee(
-        name: str = "Test E2E Licensee",
-        licensee_type: str = "Client",
-    ):
-        group = user_group_factory(user_group_account_id=licensee_account_id)
-
-        return {
-            "name": name,
-            "address": {
-                "addressLine1": "456 Licensee St",
-                "city": "Los Angeles",
-                "state": "CA",
-                "postCode": "67890",
-                "country": "US",
-            },
-            "useBuyerAddress": False,
-            "seller": {"id": seller_id},
-            "buyer": {"id": buyer_id},
-            "account": {"id": licensee_account_id},
-            "eligibility": {"client": True, "partner": False},
-            "groups": [group],
-            "type": licensee_type,
-            "status": "Enabled",
-            "defaultLanguage": "en-US",
-        }
-
-    return _licensee
