+++ tests/e2e/accounts/user_groups/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture
+def invalid_user_group_id():
+    return "UGR-0000-0000"
